% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fcm_phase.R
\name{add_fcm_phase}
\alias{add_fcm_phase}
\title{Add the food consumption matrix to the dataset}
\usage{
add_fcm_phase(
  dataset,
  fews_vars = c("fcs_cat", "rcsi_cat", "hhs_cat"),
  fcs_categories = c("Acceptable", "Poor", "Borderline"),
  rcsi_categories = c("No to Low", "Medium", "Severe"),
  hhs_categories = c("None", "Little", "Moderate", "Severe", "Very Severe")
)
}
\arguments{
\item{dataset}{A dataframe}

\item{fews_vars}{A vector of 3 strings for the column names of fcs, hhs, and rcsi categories from the dataframe
it has to include all "fcs_cat", "rcsi_cat" and "hhs_cat"}

\item{fcs_categories}{A vector of 3 strings for the values included in the fcs_cat column
It have to be a character and include "Acceptable", "Poor" and "Borderline"}

\item{rcsi_categories}{A vector of 3 strings for the values included in the rcsi_cat column
It have to be a character and include "No to Low", "Medium" and "Severe"}

\item{hhs_categories}{A vector of 5 strings for the values included in the hhs_cat column
It have to be a character and include "None","Little","Moderate"," Severe","Very Severe"}
}
\value{
this function returns a dataframe with a column called fc_cell that includes values from 1 to 45
representing the Food Consumption Score Matrix and the fc_phase column that includes the different 5 phases
of food consumption
}
\description{
Add the food consumption matrix to the dataset
}
