% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hdds.R
\name{add_hdds}
\alias{add_hdds}
\title{add_hdds}
\usage{
add_hdds(
  .dataset,
  fsl_hdds_cereals = "fsl_hdds_cereals",
  fsl_hdds_tubers = "fsl_hdds_tubers",
  fsl_hdds_veg = "fsl_hdds_veg",
  fsl_hdds_fruit = "fsl_hdds_fruit",
  fsl_hdds_meat = "fsl_hdds_meat",
  fsl_hdds_eggs = "fsl_hdds_eggs",
  fsl_hdds_fish = "fsl_hdds_fish",
  fsl_hdds_legumes = "fsl_hdds_legumes",
  fsl_hdds_dairy = "fsl_hdds_dairy",
  fsl_hdds_oil = "fsl_hdds_oil",
  fsl_hdds_sugar = "fsl_hdds_sugar",
  fsl_hdds_condiments = "fsl_hdds_condiments",
  yes_val = "yes",
  no_val = "no"
)
}
\arguments{
\item{.dataset}{the clean dataset}

\item{fsl_hdds_cereals}{the name of the variable that indicates if cereals were consumed in the last 24 hours}

\item{fsl_hdds_tubers}{the name of the variable that indicates if roots or tubers were consumed in the last 24 hours}

\item{fsl_hdds_veg}{the name of the variable that indicates if vergetables were consumed in the last 24 hours}

\item{fsl_hdds_fruit}{the name of the variable that indicates if fruits were consumed in the last 24 hours}

\item{fsl_hdds_meat}{the name of the variable that indicates if meat were consumed in the last 24 hours}

\item{fsl_hdds_eggs}{the name of the variable that indicates if eggs were consumed in the last 24 hours}

\item{fsl_hdds_fish}{the name of the variable that indicates if fish were consumed in the last 24 hours}

\item{fsl_hdds_legumes}{the name of the variable that indicates if pulses or legumes were consumed in the last 24 hours}

\item{fsl_hdds_dairy}{the name of the variable that indicates if dairy were consumed in the last 24 hours}

\item{fsl_hdds_oil}{the name of the variable that indicates if oil were consumed in the last 24 hours}

\item{fsl_hdds_sugar}{the name of the variable that indicates if sugar were consumed in the last 24 hours}

\item{fsl_hdds_condiments}{the name of the variable that indicates if condiments were consumed in the last 24 hours}

\item{yes_val}{A character value in the dataset associated with "yes"}

\item{no_val}{A character value in the dataset associated with "no"}
}
\value{
the dataset with fsl_hdds_score and fsl_hdds_cat computed, as well as
all the hdds columns recoded to 1 and 0 for yes and no as fsl_hdds_x_recoded
}
\description{
add_hdds
}
\examples{
df1 <- data.frame(fsl_hdds_cereals = c("yes", "yes", "yes", "no", "yes", "no"),
  fsl_hdds_tubers = c("yes", "yes", "yes", "no", "yes", "no"),
  fsl_hdds_veg = c("no", "yes", "yes", "no", "yes", "no"),
  fsl_hdds_fruit = c("yes", "yes", "yes", "no", "yes", "no"),
  fsl_hdds_meat = c("no", "no", "yes", "no", "yes", "no"),
  fsl_hdds_eggs = c("yes", "no", "yes", "no", "yes", "no"),
  fsl_hdds_fish = c("yes", "yes", "yes", "no", "yes", "no"),
  fsl_hdds_legumes = c("yes", "no", "yes", "no", "yes", "no"),
  fsl_hdds_dairy = c("no", "yes", "yes", "no", "yes", "no"),
  fsl_hdds_oil = c("yes", "yes", "yes", "no", "no", "no"),
  fsl_hdds_sugar = c("yes", "yes", "no", "no", "yes", "no"),
  fsl_hdds_condiments = c("no", "yes", "yes", "no", "yes", "no")
)
add_hdds(.dataset = df1
)
}
