% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hhs.R
\name{add_hhs}
\alias{add_hhs}
\title{Add the household hunger scale to the dataset}
\usage{
add_hhs(
  .dataset,
  fsl_hhs_nofoodhh = "fsl_hhs_nofoodhh",
  fsl_hhs_nofoodhh_freq = "fsl_hhs_nofoodhh_freq",
  fsl_hhs_sleephungry = "fsl_hhs_sleephungry",
  fsl_hhs_sleephungry_freq = "fsl_hhs_sleephungry_freq",
  fsl_hhs_alldaynight = "fsl_hhs_alldaynight",
  fsl_hhs_alldaynight_freq = "fsl_hhs_alldaynight_freq",
  yes_answer = "yes",
  no_answer = "no",
  rarely_answer = "rarely",
  sometimes_answer = "sometimes",
  often_answer = "often"
)
}
\arguments{
\item{.dataset}{Dataset}

\item{fsl_hhs_nofoodhh}{The name of the column "In the past 4 weeks (30 days), was there ever no food to eat
of any kind in your house because of lack of resources to get food?". It has to be a string.}

\item{fsl_hhs_nofoodhh_freq}{The name of the column "How often did this happen in the past (4 weeks/30 days)?".
It has to be a string.}

\item{fsl_hhs_sleephungry}{The name of the column "In the past 4 weeks (30 days), did you or any household
member go to sleep at night hungry because there was not enough food?". It has to be a string.}

\item{fsl_hhs_sleephungry_freq}{The name of the column "How often did this happen in the past (4 weeks/30 days)?".
It has to be a string.}

\item{fsl_hhs_alldaynight}{The name of the column "In the past 4 weeks (30 days), did you or any household
member go a whole day and night without eating anything at all because there was not enough food?".
It has to be a string.}

\item{fsl_hhs_alldaynight_freq}{The name of the column "How often did this happen in the past (4 weeks/30 days)?".
It has to be a string.}

\item{yes_answer}{Value used for "yes"}

\item{no_answer}{Value used for the "no"}

\item{rarely_answer}{Value used for "rarely"}

\item{sometimes_answer}{Value used for "sometimes"}

\item{often_answer}{Value used for "often"}
}
\value{
It returns the dataframe with 12 extras columns: recoded hhs questions, score for the 3 sets of questions
(from 0 to 2), the HHS score (from 0 to 6), the HHS category and the HHS IPC category
}
\description{
Add the household hunger scale to the dataset
}
\examples{
{
  input_data <- data.frame(
    fsl_hhs_nofoodhh = c("no", "yes", "no", "no", "no"),
    fsl_hhs_nofoodhh_freq = c(NA_character_, "rarely",
     NA_character_, NA_character_, NA_character_),
    fsl_hhs_sleephungry = c("no", "no", "yes", "no", "no"),
    fsl_hhs_sleephungry_freq = c(NA_character_, NA_character_,
     "often", NA_character_, NA_character_),
    fsl_hhs_alldaynight = c("no", "no", "yes", "yes", "yes"),
    fsl_hhs_alldaynight_freq = c(NA_character_, NA_character_,
     "often", "rarely", "sometimes")
  )

  add_hhs(
    .dataset = input_data,
    fsl_hhs_nofoodhh = "fsl_hhs_nofoodhh",
    fsl_hhs_nofoodhh_freq = "fsl_hhs_nofoodhh_freq",
    fsl_hhs_sleephungry = "fsl_hhs_sleephungry",
    fsl_hhs_sleephungry_freq = "fsl_hhs_sleephungry_freq",
    fsl_hhs_alldaynight = "fsl_hhs_alldaynight",
    fsl_hhs_alldaynight_freq = "fsl_hhs_alldaynight_freq",
    yes_answer = "yes",
    no_answer = "no",
    rarely_answer = "rarely",
    sometimes_answer = "sometimes",
    often_answer = "often"
  )
}
}
