% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rcsi.R
\name{add_rcsi}
\alias{add_rcsi}
\title{Add indicator for reduced Household CSI Score(rcsi)}
\usage{
add_rcsi(
  .dataset,
  fsl_rcsi_lessquality = "fsl_rcsi_lessquality",
  fsl_rcsi_borrow = "fsl_rcsi_borrow",
  fsl_rcsi_mealsize = "fsl_rcsi_mealsize",
  fsl_rcsi_mealadult = "fsl_rcsi_mealadult",
  fsl_rcsi_mealnb = "fsl_rcsi_mealnb"
)
}
\arguments{
\item{.dataset}{dataset}

\item{fsl_rcsi_lessquality}{Column representing question- During the last 7 days, were there days (and, if so, how many) when your household had to rely on less preferred and less expensive food to cope with a lack of food or money to buy it?}

\item{fsl_rcsi_borrow}{Column representing question- During the last 7 days, were there days (and, if so, how many) when your household had to borrow food or rely on help from a relative or friend to cope with a lack of food or money to buy it?}

\item{fsl_rcsi_mealsize}{Column representing question- During the last 7 days, were there days (and, if so, how many) when your household had to limit portion size of meals at meal times to cope with a lack of food or money to buy it?}

\item{fsl_rcsi_mealadult}{Column representing question- During the last 7 days, were there days (and, if so, how many) when your household had to restrict consumption by adults in order for small children to eat to cope with a lack of food or money to buy it?}

\item{fsl_rcsi_mealnb}{Column representing question - During the last 7 days, were there days (and, if so, how many) when your household had to reduce number of meals eaten in a day to cope with a lack of food or money to buy it?}
}
\value{
A dataset with one additional column.
}
\description{
Add indicator for reduced Household CSI Score(rcsi)
}
\examples{
test_data <- data.frame(
  fsl_rcsi_lessquality = c(1, 2, 3, 1),
  fsl_rcsi_borrow = c(0, 0, 3, 0),
  fsl_rcsi_mealsize = c(4, 2, 6, 1),
  fsl_rcsi_mealadult = c(4, 3, 5, 0),
  fsl_rcsi_mealnb = c(2, 5, NA_integer_, 1)
)
add_rcsi(test_data)
}
