% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addindicators-package.R
\docType{package}
\name{addindicators-package}
\alias{addindicators}
\alias{addindicators-package}
\title{addindicators: The addindicators Package Focuses on Adding Indicators to a Dataset and Reviews the Added Indicator}
\description{
The addindicators package focuses on adding indicators such as Food Consumption Score (FCS), Household Hunger Score(HHS) etc to a dataset and reviews the added indicator.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://impact-initiatives.github.io/addindicators/}
}

}
\author{
\strong{Maintainer}: Mehedi Khan \email{mh.khan@reach-initiative.org} (\href{https://orcid.org/0000-0002-4276-4485}{ORCID})

Authors:
\itemize{
  \item Yann Say \email{yann.say@impact-initiatives.org} (\href{https://orcid.org/0000-0002-7390-4209}{ORCID})
}

}
\keyword{internal}
