% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{addindicators_analysis_by_group}
\alias{addindicators_analysis_by_group}
\title{Analysis by population group}
\usage{
addindicators_analysis_by_group
}
\description{
Survey data
}
\details{
These data sets include HH data (raw and clean data) and analysis along with cleaning log.
}
\examples{
addindicators_analysis_by_group
addindicators_raw_data
addindicators_cleaning_log
addindicators_clean_data
addindicators_survey
addindicators_choices
addindicators_overall_analysis
}
\keyword{datasets}
