% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{addindicators_MSNA_template_data}
\alias{addindicators_MSNA_template_data}
\title{MSNA template dataset (example)}
\usage{
addindicators_MSNA_template_data
}
\description{
MSNA template dataset (example)
}
\keyword{datasets}
