% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hhs.R
\name{add_hhs}
\alias{add_hhs}
\title{Add the household hunger scale to the dataset}
\usage{
add_hhs(
  .dataset,
  hhs_nofoodhh_1 = "fs_hhs_nofood_yn",
  hhs_nofoodhh_1a = "fs_hhs_nofood_freq",
  hhs_sleephungry_2 = "fs_hhs_sleephungry_yn",
  hhs_sleephungry_2a = "fs_hhs_sleephungry_freq",
  hhs_alldaynight_3 = "fs_hhs_daynoteating_yn",
  hhs_alldaynight_3a = "fs_hhs_daynoteating_freq",
  yes_answer = "yes",
  no_answer = "no",
  rarely_answer = "rarely_1_2",
  sometimes_answer = "sometimes_3_10",
  often_answer = "often_10_times"
)
}
\arguments{
\item{.dataset}{Dataset}

\item{hhs_nofoodhh_1}{The name of the column "In the past 4 weeks (30 days), was there ever no food to eat
of any kind in your house because of lack of resources to get food?". It has to be a string.}

\item{hhs_nofoodhh_1a}{The name of the column "How often did this happen in the past (4 weeks/30 days)?".
It has to be a string.}

\item{hhs_sleephungry_2}{The name of the column "In the past 4 weeks (30 days), did you or any household
member go to sleep at night hungry because there was not enough food?". It has to be a string.}

\item{hhs_sleephungry_2a}{The name of the column "How often did this happen in the past (4 weeks/30 days)?".
It has to be a string.}

\item{hhs_alldaynight_3}{The name of the column "In the past 4 weeks (30 days), did you or any household
member go a whole day and night without eating anything at all because there was not enough food?".
It has to be a string.}

\item{hhs_alldaynight_3a}{The name of the column "How often did this happen in the past (4 weeks/30 days)?".
It has to be a string.}

\item{yes_answer}{Value used for "Yes"}

\item{no_answer}{Value used for the "No"}

\item{rarely_answer}{Value used for "Rarely (1-2)"}

\item{sometimes_answer}{Value used for "Sometimes (3-10)"}

\item{often_answer}{Value used for "Often (10+ times)}
}
\value{
It returns the dataframe with 12 extras columns: recoded hhs questions, score for the 3 sets of questions
(from 0 to 2), the HHS score (from 0 to 6), the HHS category and the HHS IPC category
}
\description{
Add the household hunger scale to the dataset
}
\examples{
{
  input_data <- data.frame(
    fs_hhs_nofood_yn = c("no", "yes", "no", "no", "no"),
    fs_hhs_nofood_freq = c(NA_character_, "rarely_1_2", NA_character_, NA_character_, NA_character_),
    fs_hhs_sleephungry_yn = c("no", "no", "yes", "no", "no"),
    fs_hhs_sleephungry_freq = c(NA_character_, NA_character_, "often_10_times", NA_character_, NA_character_),
    fs_hhs_daynoteating_yn = c("no", "no", "yes", "yes", "yes"),
    fs_hhs_daynoteating_freq = c(NA_character_, NA_character_, "often_10_times", "rarely_1_2", "sometimes_3_10")
  )

  add_hhs(
    .dataset = input_data,
    hhs_nofoodhh_1 = "fs_hhs_nofood_yn",
    hhs_nofoodhh_1a = "fs_hhs_nofood_freq",
    hhs_sleephungry_2 = "fs_hhs_sleephungry_yn",
    hhs_sleephungry_2a = "fs_hhs_sleephungry_freq",
    hhs_alldaynight_3 = "fs_hhs_daynoteating_yn",
    hhs_alldaynight_3a = "fs_hhs_daynoteating_freq",
    yes_answer = "yes",
    no_answer = "no",
    rarely_answer = "rarely_1_2",
    sometimes_answer = "sometimes_3_10",
    often_answer = "often_10_times"
  )
}
}
