% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lcsi.R
\name{add_lcsi}
\alias{add_lcsi}
\title{Add LCSI}
\usage{
add_lcsi(
  .dataset,
  lcsi_stress_vars,
  lcsi_crisis_vars,
  lcsi_emergency_vars,
  yes_val = NULL,
  no_val = NULL,
  exhausted_val = NULL,
  not_applicable_val = NULL
)
}
\arguments{
\item{.dataset}{A dataframe with the ten LCSI variables needed for analysis.}

\item{lcsi_stress_vars}{A vector of character values that are the column names for the four stress LCSI variables.}

\item{lcsi_crisis_vars}{A vector of character values that are the column names for the three crisis LCSI variables.}

\item{lcsi_emergency_vars}{A vector of character values that are the column names for the thre emergency LCSI variables.}

\item{yes_val}{A character value in the dataset associated with "Yes, used this coping strategy in the last 30 days."}

\item{no_val}{A character value in the dataset associated with "No, have not used this coping strategy in the last 30 days."}

\item{exhausted_val}{A character value in the dataset associated with "No, haven't used in the last 30 days because I've exhausted this coping strategy in the last 6 or 12 months."}

\item{not_applicable_val}{A character value in the dataset associated with "This coping strategy is not applicable for the household.}
}
\value{
Returns a dataframe with added columns for LCSI indicators.
}
\description{
Function to calculate Livelihood Coping Strategy Index (LCSI)
}
\examples{
{
  input_data1 <- data.frame(
stress1 =             c("No", "No", "Exhausted", "Not Applicable", "No"),
stress2 =             c("No", "Yes", "Not Applicable", "No", "No"),
stress3 =             c("Not Applicable", "Not Applicable", "Yes", "No", "No"),
stress4 =             c("Not Applicable", "No", "Yes", "Yes", "No"),
crisis1 =             c("No", "Not Applicable", "Yes", "Exhausted", "No"),
crisis2 =             c("No", "No", "No", "No", "No"),
crisis3 =             c("No", "No", "Yes", "Not Applicable", "No"),
emergency1 =          c("No", "Not Applicable", "Not Applicable", "No", "No"),
emergency2 =          c("No", "Not Applicable", "Yes", "Not Applicable", "No"),
emergency3 =          c("Not Applicable", "No", "Not Applicable", "No", "Exhausted"))

add_lcsi(.dataset = input_data1,
lcsi_stress_vars = c("stress1", "stress2", "stress3", "stress4"),
lcsi_crisis_vars = c("crisis1", "crisis2", "crisis3"),
lcsi_emergency_vars = c("emergency1", "emergency2", "emergency3"),
yes_val = "Yes",
no_val = "No",
exhausted_val = "Exhausted",
not_applicable_val = "Not Applicable")

}
}
