% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rcsi.R
\name{add_rcsi}
\alias{add_rcsi}
\title{Add indicator for reduced Household CSI Score(rcsi)}
\usage{
add_rcsi(
  data,
  rCSILessQlty = "rCSILessQlty",
  rCSIBorrow = "rCSIBorrow",
  rCSIMealSize = "rCSIMealSize",
  rCSIMealAdult = "rCSIMealAdult",
  rCSIMealNb = "rCSIMealNb",
  new_colname = "rcsi"
)
}
\arguments{
\item{data}{dataset}

\item{rCSILessQlty}{Column representing question- During the last 7 days, were there days (and, if so, how many) when your household had to rely on less preferred and less expensive food to cope with a lack of food or money to buy it?}

\item{rCSIBorrow}{Column representing question- During the last 7 days, were there days (and, if so, how many) when your household had to borrow food or rely on help from a relative or friend to cope with a lack of food or money to buy it?}

\item{rCSIMealSize}{Column representing question- During the last 7 days, were there days (and, if so, how many) when your household had to limit portion size of meals at meal times to cope with a lack of food or money to buy it?}

\item{rCSIMealAdult}{Column representing question- During the last 7 days, were there days (and, if so, how many) when your household had to restrict consumption by adults in order for small children to eat to cope with a lack of food or money to buy it?}

\item{rCSIMealNb}{Column representing question - During the last 7 days, were there days (and, if so, how many) when your household had to reduce number of meals eaten in a day to cope with a lack of food or money to buy it?}

\item{new_colname}{The prefix for the new columns. It has to be a string.}
}
\value{
A dataset with one additional column.
}
\description{
Add indicator for reduced Household CSI Score(rcsi)
}
\examples{
test_data <- data.frame(
rCSILessQlty = c(1,2,3,1),
rCSIBorrow = c(0,0,3,0),
rCSIMealSize = c(4,2,6,1),
rCSIMealAdult = c(4,3,5,0),
rCSIMealNb = c(2,5,NA_integer_,1)
)
add_rcsi(test_data)
}
